refactor: 강의 기타정보를 담당하는 LectureInfoRepository 신설
