+++ app/src/main/java/com/wafflestudio/snutt2/data/lecture_search/LectureSearchRepository.kt
@@ -1,13 +1,9 @@
 package com.wafflestudio.snutt2.data.lecture_search
 
 import androidx.paging.PagingData
-import com.wafflestudio.snutt2.domainmodel.CourseBook
-import com.wafflestudio.snutt2.domainmodel.LectureReviewInfo
 import com.wafflestudio.snutt2.domainmodel.SearchTag
 import com.wafflestudio.snutt2.domainmodel.SearchTime
 import com.wafflestudio.snutt2.domainmodel.SearchedLecture
-import com.wafflestudio.snutt2.lib.network.Result
-import com.wafflestudio.snutt2.lib.network.dto.core.LectureBuildingDto
 import kotlinx.coroutines.flow.Flow
 
 interface LectureSearchRepository {
@@ -28,10 +24,4 @@ interface LectureSearchRepository {
     fun storeRecentSearchedDepartment(tag: SearchTag)
 
     fun removeRecentSearchedDepartment(tag: SearchTag)
-
-    suspend fun getBuildings(places: List<String>): Result<List<LectureBuildingDto>>
-
-    suspend fun getSyllabusUrl(courseBook: CourseBook, courseNumber: String, lectureNumber: String): Result<String>
-
-    suspend fun getReviewInfo(lectureId: String): Result<LectureReviewInfo?>
 }
